+++ harvester/encoding.go
@@ -8,41 +8,54 @@ import (
 	"golang.org/x/text/encoding/charmap"
 	"golang.org/x/text/encoding/htmlindex"
 	"golang.org/x/text/encoding/simplifiedchinese"
-	"golang.org/x/text/encoding/unicode"
 	"golang.org/x/text/transform"
 )
 
 // Decoder wraps a reader for decoding input to utf-8 on read.
 type Decoder func(io.Reader) io.Reader
 
-var encodings = map[string]Decoder{
+var encodings = map[string]encoding.Encoding{
 	// default
 	"nop":   Plain,
 	"plain": Plain,
 
 	// utf8 (validate input) - shadow htmlindex utf8 codecs not validating input
-	"unicode-1-1-utf-8": trans(encoding.UTF8Validator),
-	"utf-8":             trans(encoding.UTF8Validator),
-	"utf8":              trans(encoding.UTF8Validator),
+	"unicode-1-1-utf-8": utf8Encoding{},
+	"utf-8":             utf8Encoding{},
+	"utf8":              utf8Encoding{},
 
 	// utf16
-	"utf-16be-bom": enc(unicode.UTF16(unicode.BigEndian, unicode.UseBOM)),
+	// "utf-16be-bom": unicode.UTF16(unicode.BigEndian, unicode.UseBOM),
 
 	// simplified chinese
-	"gbk": enc(simplifiedchinese.GBK), // shadow htmlindex using 'GB10830' for GBK
+	"gbk": simplifiedchinese.GBK, // shadow htmlindex using 'GB10830' for GBK
 
 	// 8bit charmap encodings
-	"iso8859-6e": enc(charmap.ISO8859_6E),
-	"iso8859-6i": enc(charmap.ISO8859_6I),
-	"iso8859-8e": enc(charmap.ISO8859_8E),
-	"iso8859-8i": enc(charmap.ISO8859_8I),
+	"iso8859-6e": charmap.ISO8859_6E,
+	"iso8859-6i": charmap.ISO8859_6I,
+	"iso8859-8e": charmap.ISO8859_8E,
+	"iso8859-8i": charmap.ISO8859_8I,
 }
 
 // Plain file encoding not transforming any read bytes.
-var Plain = nopEnc
+var Plain = encoding.Nop
+
+// UTF-8 encoding copying input to output sequence replacing invalid UTF-8
+// converted to '\uFFFD'.
+//
+// See: http://encoding.spec.whatwg.org/#replacement
+type utf8Encoding struct{}
+
+func (utf8Encoding) NewDecoder() transform.Transformer {
+	return encoding.Replacement.NewEncoder()
+}
+
+func (utf8Encoding) NewEncoder() transform.Transformer {
+	return encoding.Replacement.NewEncoder()
+}
 
 // Find returns
-func findEncoding(name string) (Decoder, bool) {
+func findEncoding(name string) (encoding.Encoding, bool) {
 	if name == "" {
 		return Plain, true
 	}
@@ -52,20 +65,5 @@ func findEncoding(name string) (Decoder, bool) {
 	}
 
 	codec, err := htmlindex.Get(name)
-	if err != nil {
-		return nil, false
-	}
-	return enc(codec), true
-}
-
-func nopEnc(r io.Reader) io.Reader { return r }
-
-func enc(encoding encoding.Encoding) Decoder {
-	return trans(encoding.NewDecoder())
-}
-
-func trans(t transform.Transformer) Decoder {
-	return func(r io.Reader) io.Reader {
-		return transform.NewReader(r, t)
-	}
+	return codec, err == nil
 }