+++ CHANGELOG.md
@@ -6,10 +6,13 @@ All notable changes to this project will be documented in this file based on the
 ## [Unreleased](https://github.com/elastic/libbeat/compare/1.0.0-rc1...HEAD)
 
 ### Backward Compatibility Breaks
+- Removed utf-16be-bom encoding support. Support will be added with fix for #205
 
 ### Bugfixes
 - Filebeat will now exit if a configuration error is detected. #198
 - Fix to enable prospector to harvest existing files that are modified. #199
+- Improve line reading and encoding to better keep track of file offsets based
+  on encoding. #224
 
 ### Added